keep the old tool and add deprecation warning
