+++ src/plopp/widgets/cut3d.py
@@ -1,73 +1,19 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
 
-import asyncio
-from functools import reduce
-from typing import Any, Callable, Dict, Literal, Tuple
+import warnings
+from functools import partial
+from typing import Any, Dict, Literal, Tuple
 
 import ipywidgets as ipw
 import numpy as np
 import scipp as sc
 
-from ..core import Node
-from ..graphics import BaseFig
+from ..core import View, node
+from ..core.typing import VisibleDeprecationWarning
+from .debounce import debounce
 from .style import BUTTON_LAYOUT
-
-OPERATIONS = {'or': sc.logical_or, 'and': sc.logical_and, 'xor': sc.logical_xor}
-
-
-class Timer:
-    """
-    From:
-    https://ipywidgets.readthedocs.io/en/8.0.2/examples/Widget%20Events.html#Debouncing
-    """
-
-    def __init__(self, timeout: float, callback: Callable):
-        self._timeout = timeout
-        self._callback = callback
-
-    async def _job(self):
-        await asyncio.sleep(self._timeout)
-        self._callback()
-
-    def start(self):
-        self._task = asyncio.ensure_future(self._job())
-
-    def cancel(self):
-        self._task.cancel()
-
-
-def debounce(wait: float):
-    """
-    Decorator that will postpone a function's
-    execution until after `wait` seconds
-    have elapsed since the last time it was invoked.
-
-    From:
-    https://ipywidgets.readthedocs.io/en/8.0.2/examples/Widget%20Events.html#Debouncing
-    """
-
-    def decorator(fn: Callable):
-        timer = None
-
-        def debounced(*args, **kwargs):
-            nonlocal timer
-
-            def call_it():
-                fn(*args, **kwargs)
-
-            if timer is not None:
-                timer.cancel()
-            timer = Timer(wait, call_it)
-            timer.start()
-
-        return debounced
-
-    return decorator
-
-
-def select(da: sc.DataArray, s: Tuple[str, sc.Variable]) -> sc.DataArray:
-    return da[s]
+from .tools import PlusMinusTool
 
 
 class Cut3dTool(ipw.HBox):
@@ -82,137 +28,196 @@ class Cut3dTool(ipw.HBox):
 
     Parameters
     ----------
+    view:
+        The 3d figure that contains the point clouds to be cut.
     limits:
         The spatial extent of the points in the 3d figure in the XYZ directions.
     direction:
         The direction normal to the slice.
-    plain_update:
-        A function to update the scene without debouncing.
-    throttled_update:
-        A function to update the scene with debouncing.
+    value:
+        Set the cut to active upon creation if ``True``.
     color:
         Color of the cut's outline.
     linewidth:
         Width of the line delineating the outline.
+    **kwargs:
+        The kwargs are forwarded to the ToggleButton from ``ipywidgets``.
     """
 
     def __init__(
         self,
+        view: View,
         limits: Tuple[sc.Variable, sc.Variable, sc.Variable],
         direction: Literal['x', 'y', 'z'],
-        plain_update: Callable,
-        throttled_update: Callable,
+        value: bool = False,
         color: str = 'red',
         linewidth: float = 1.5,
+        **kwargs,
     ):
         import pythreejs as p3
 
         self._limits = limits
         self._direction = direction
         axis = 'xyz'.index(self._direction)
-        self.dim = self._limits[axis].dim
+        self._dim = self._limits[axis].dim
         self._unit = self._limits[axis].unit
-        self.visible = True
-        self._plain_update = plain_update
-        self._throttled_update = throttled_update
+        self._view = view
 
         w_axis = 2 if self._direction == 'x' else 0
         h_axis = 2 if self._direction == 'y' else 1
         width = (self._limits[w_axis][1] - self._limits[w_axis][0]).value
         height = (self._limits[h_axis][1] - self._limits[h_axis][0]).value
 
-        self.outlines = [
-            p3.LineSegments(
-                geometry=p3.EdgesGeometry(
-                    p3.PlaneBufferGeometry(width=width, height=height)
-                ),
-                material=p3.LineBasicMaterial(color=color, linewidth=linewidth),
+        self.outline = p3.LineSegments(
+            geometry=p3.EdgesGeometry(
+                p3.PlaneBufferGeometry(width=width, height=height)
             ),
-            p3.LineSegments(
-                geometry=p3.EdgesGeometry(
-                    p3.PlaneBufferGeometry(width=width, height=height)
-                ),
-                material=p3.LineBasicMaterial(color=color, linewidth=linewidth),
-            ),
-        ]
+            material=p3.LineBasicMaterial(color=color, linewidth=linewidth),
+        )
         if self._direction == 'x':
-            for outline in self.outlines:
-                outline.rotateY(0.5 * np.pi)
+            self.outline.rotateY(0.5 * np.pi)
         if self._direction == 'y':
-            for outline in self.outlines:
-                outline.rotateX(0.5 * np.pi)
+            self.outline.rotateX(0.5 * np.pi)
 
         center = [var.mean().value for var in self._limits]
-        vmin = self._limits[axis][0].value
-        vmax = self._limits[axis][1].value
-        dx = vmax - vmin
-        delta = 0.05 * dx
-        self.slider = ipw.FloatRangeSlider(
-            min=vmin,
-            max=vmax,
-            value=[center[axis] - delta, center[axis] + delta],
-            step=dx * 0.01,
-            description=direction.upper(),
-            style={'description_width': 'initial'},
-            layout={'width': '470px', 'padding': '0px'},
+        self.outline.position = center
+        self.outline.visible = value
+
+        self.thickness = 0.01
+        self._thickness_step = self.thickness * 0.25
+
+        self.button = ipw.ToggleButton(value=value, **kwargs)
+        self.slider = ipw.FloatSlider(
+            min=limits[axis][0].value,
+            max=limits[axis][1].value,
+            value=center[axis],
+            layout={'width': '150px', 'padding': '0px'},
+            disabled=not value,
+            readout=False,
         )
-
-        self.cut_visible = ipw.Button(
-            icon='eye-slash',
-            tooltip='Hide cut',
-            layout={'width': '16px', 'padding': '0px'},
+        self.slider.step = (self.slider.max - self.slider.min) * self.thickness * 0.5
+        self.readout = ipw.FloatText(
+            layout={'width': '70px'}, step=self.slider.step, disabled=not value
         )
-
-        for outline, val in zip(self.outlines, self.slider.value):
-            pos = list(center)
-            pos[axis] = val
-            outline.position = pos
-
         self.unit_label = ipw.Label(f'[{self._unit}]')
-        self.cut_visible.on_click(self.toggle)
+        ipw.jslink((self.slider, "value"), (self.readout, "value"))
+
+        self.plus_minus = PlusMinusTool(
+            plus={
+                'callback': self.increase_thickness,
+                'tooltip': 'Increase cut thickness',
+                'disabled': not value,
+            },
+            minus={
+                'callback': self.decrease_thickness,
+                'tooltip': 'Decrease cut thickness',
+                'disabled': not value,
+            },
+        )
+
+        self.button.observe(self.toggle, names='value')
         self.slider.observe(self.move, names='value')
+        self.slider.observe(self.update_cut, names='value')
 
-        super().__init__([self.slider, ipw.Label(f'[{self._unit}]'), self.cut_visible])
+        self._nodes = list(self._view.graph_nodes.values())
+        self.select_nodes = {}
 
-    def toggle(self, owner):
-        """
-        Toggle the visibility of the cut on and off.
-        """
-        self.visible = not self.visible
-        for outline in self.outlines:
-            outline.visible = self.visible
-        self.slider.disabled = not self.visible
-        owner.icon = 'eye-slash' if self.visible else 'eye'
-        owner.tooltip = 'Hide cut' if self.visible else 'Show cut'
-        self._plain_update()
+        super().__init__(
+            [
+                ipw.VBox([self.button, self.plus_minus]),
+                ipw.VBox(
+                    [self.slider, ipw.HBox([self.readout, self.unit_label])],
+                    layout={'align_items': 'center'},
+                ),
+            ]
+        )
 
-    def toggle_border(self, value: bool):
+    def toggle(self, change: Dict[str, Any]):
         """
-        Toggle the border visbility.
+        Toggle the tool on and off.
         """
-        for outline in self.outlines:
-            outline.visible = value
+        self.outline.visible = change['new']
+        for widget in (self.slider, self.plus_minus, self.readout):
+            widget.disabled = not change['new']
+        if change['new']:
+            self._add_cut()
+        else:
+            self._remove_cut()
 
     def move(self, value: Dict[str, Any]):
         """
         Move the outline of the cut according to new position given by the slider.
         """
-        # Early return if relative difference between new and old value is small
-        old = np.array(value['old'])
-        if np.abs((np.array(value['new']) - old) / old).max() < 1e-6:
-            return
-        for outline, val in zip(self.outlines, value['new']):
-            pos = list(outline.position)
-            axis = 'xyz'.index(self._direction)
-            pos[axis] = val
-            outline.position = pos
-        self._throttled_update()
+        pos = list(self.outline.position)
+        axis = 'xyz'.index(self._direction)
+        pos[axis] = value['new']
+        self.outline.position = pos
+        self._remove_cut()
+
+    def _add_cut(self):
+        """
+        Add a point cloud representing the points that intersect the cut plane.
+        """
+
+        def select(da, s):
+            return da[s]
+
+        for n in self._nodes:
+            da = n.request_data()
+            delta = sc.scalar(
+                (self.slider.max - self.slider.min) * self.thickness, unit=self._unit
+            )
+            pos = sc.scalar(self.slider.value, unit=self._unit)
+            selection = sc.abs(da.coords[self._dim] - pos) < delta
+            if selection.sum().value > 0:
+                self.select_nodes[n.id] = node(partial(select, s=selection))(da=n)
+                self.select_nodes[n.id].add_view(self._view)
+                self._view.update(
+                    self.select_nodes[n.id].request_data(),
+                    key=self.select_nodes[n.id].id,
+                )
+
+    def _remove_cut(self):
+        """
+        Remove a cut point the scene.
+        """
+        for n in self.select_nodes.values():
+            self._view.remove(n.id)
+            n.remove()
+        self.select_nodes.clear()
+
+    @debounce(0.3)
+    def update_cut(self, _):
+        """
+        Update the position of the point cloud. This uses the debounce mechanism to
+        avoid updating the cloud too often, and instead rely on simply moving the
+        outline, which is cheap.
+        """
+        if self.value:
+            self._add_cut()
 
     @property
-    def range(self):
-        return sc.scalar(self.slider.value[0], unit=self._unit), sc.scalar(
-            self.slider.value[1], unit=self._unit
-        )
+    def value(self):
+        """
+        Return the state of the ToggleButton.
+        """
+        return self.button.value
+
+    def increase_thickness(self):
+        """
+        Increase the thickness of the cut.
+        """
+        self.thickness = self.thickness + self._thickness_step
+        self._remove_cut()
+        self._add_cut()
+
+    def decrease_thickness(self):
+        """
+        Decrease the thickness of the cut.
+        """
+        self.thickness = max(self.thickness - self._thickness_step, 0)
+        self._remove_cut()
+        self._add_cut()
 
 
 class TriCutTool(ipw.HBox):
@@ -226,39 +231,43 @@ class TriCutTool(ipw.HBox):
         The 3d figure that contains the point clouds to be cut.
     """
 
-    def __init__(self, fig: BaseFig):
-        self._view = fig._view
-        self._limits = self._view.get_limits()
-        self.cuts = []
-        self._operation = 'or'
+    def __init__(self, fig: View):
+        warnings.warn(
+            "This 3D cut tool is deprecated and will be removed in the future. "
+            "Use the new `ClippingPlanes` tool instead.",
+            VisibleDeprecationWarning,
+        )
 
-        self.tabs = ipw.Tab(layout={'width': '550px'})
-        self._original_nodes = list(self._view.graph_nodes.values())
-        self._nodes = {}
-
-        self.add_cut_label = ipw.Label('Add cut:')
-        layout = {'width': '45px', 'padding': '0px 0px 0px 0px'}
-        self.add_x_cut = ipw.Button(
+        self._fig = fig
+        limits = self._fig.get_limits()
+        self.cut_x = Cut3dTool(
+            view=self._fig,
+            direction='x',
+            limits=limits,
             description='X',
-            icon='plus',
-            tooltip='Add X cut',
-            layout=layout,
+            icon='cube',
+            **BUTTON_LAYOUT,
         )
-        self.add_y_cut = ipw.Button(
+        self.cut_y = Cut3dTool(
+            view=self._fig,
+            direction='y',
+            limits=limits,
             description='Y',
-            icon='plus',
-            tooltip='Add Y cut',
-            layout=layout,
+            icon='cube',
+            **BUTTON_LAYOUT,
         )
-        self.add_z_cut = ipw.Button(
+        self.cut_z = Cut3dTool(
+            view=self._fig,
+            direction='z',
+            limits=limits,
             description='Z',
-            icon='plus',
-            tooltip='Add Z cut',
-            layout=layout,
+            icon='cube',
+            **BUTTON_LAYOUT,
         )
-        self.add_x_cut.on_click(lambda _: self._add_cut('x'))
-        self.add_y_cut.on_click(lambda _: self._add_cut('y'))
-        self.add_z_cut.on_click(lambda _: self._add_cut('z'))
+
+        self._fig.canvas.add(
+            [self.cut_x.outline, self.cut_y.outline, self.cut_z.outline]
+        )
 
         self.opacity = ipw.BoundedFloatText(
             min=0,
@@ -266,169 +275,47 @@ class TriCutTool(ipw.HBox):
             step=0.01,
             disabled=True,
             value=0.03,
-            description='Opacity:',
-            tooltip='Set the opacity of the background',
             style={'description_width': 'initial'},
-            layout={'width': '142px', 'padding': '0px 0px 0px 0px'},
+            layout={'width': '50px', 'padding': '0px 0px 0px 0px'},
         )
         self.opacity.observe(self._set_opacity, names='value')
 
-        self.cut_borders_visibility = ipw.ToggleButton(
-            value=True,
-            disabled=True,
-            icon='border-style',
-            tooltip='Toggle visbility of the borders of the cuts',
-            **BUTTON_LAYOUT,
-        )
-        self.cut_borders_visibility.observe(
-            self.toggle_border_visibility, names='value'
-        )
+        self.cut_x.button.observe(self._toggle_opacity, names='value')
+        self.cut_y.button.observe(self._toggle_opacity, names='value')
+        self.cut_z.button.observe(self._toggle_opacity, names='value')
 
-        self.cut_operation = ipw.Dropdown(
-            options=['OR', 'AND', 'XOR'],
-            value='OR',
-            disabled=True,
-            tooltip='Operation to combine multiple cuts',
-            layout={'width': '60px', 'padding': '0px 0px 0px 0px'},
-        )
-        self.cut_operation.observe(self.change_operation, names='value')
-
-        self.delete_cut = ipw.Button(
-            tooltip='Delete cut',
-            icon='trash',
-            disabled=True,
-            **BUTTON_LAYOUT,
-        )
-        self.delete_cut.on_click(self._remove_cut)
-
+        space = ipw.HBox([], layout={'width': '10px'})
         super().__init__(
             [
-                self.tabs,
-                ipw.VBox(
-                    [
-                        ipw.HBox([self.add_x_cut, self.add_y_cut, self.add_z_cut]),
-                        self.opacity,
-                        ipw.HBox(
-                            [
-                                self.cut_borders_visibility,
-                                self.cut_operation,
-                                self.delete_cut,
-                            ]
-                        ),
-                    ]
-                ),
+                self.cut_x,
+                space,
+                self.cut_y,
+                space,
+                self.cut_z,
+                space,
+                ipw.VBox([ipw.Label('Opacity:'), self.opacity]),
             ]
         )
-
         self.layout.display = 'none'
 
-    def _add_cut(self, direction: Literal['x', 'y', 'z']):
+    def _toggle_opacity(self, _):
         """
-        Add a cut in the specified direction.
+        If any cut is active, set the opacity of the original children (not the cuts) to
+        a low value. If all cuts are inactive, set the opacity back to 1.
         """
-        cut = Cut3dTool(
-            direction=direction,
-            limits=self._limits,
-            plain_update=self.update_state,
-            throttled_update=self.throttled_update,
-        )
-        self._view.canvas.add(cut.outlines)
-        self.cuts.append(cut)
-        self.tabs.children = list(self.tabs.children) + [cut]
-        self.tabs.selected_index = len(self.cuts) - 1
-        self.update_controls()
-        self.update_state()
-
-    def _remove_cut(self, _):
-        cut = self.cuts.pop(self.tabs.selected_index)
-        self._view.canvas.remove(cut.outlines)
-        self.tabs.children = self.cuts
-        self.update_state()
-        self.update_controls()
-
-    def update_controls(self):
-        """
-        If there are no active cuts, disable the controls.
-        If there is at least one cut, set the opacity of the original children
-        (not the cuts) to a low value, otherwise set it back to 1.
-        """
-        at_least_one_cut = any(1 for cut in self.cuts if cut.visible)
-        self.delete_cut.disabled = not at_least_one_cut
-        self.cut_borders_visibility.disabled = not at_least_one_cut
-        self.cut_operation.disabled = not at_least_one_cut
-        self.tabs.titles = [cut._direction.upper() for cut in self.cuts]
-        self.opacity.disabled = not at_least_one_cut
-        opacity = self.opacity.value if at_least_one_cut else 1.0
+        active_cut = any([self.cut_x.value, self.cut_y.value, self.cut_z.value])
+        self.opacity.disabled = not active_cut
+        opacity = self.opacity.value if active_cut else 1.0
         self._set_opacity({'new': opacity})
 
     def _set_opacity(self, change: Dict[str, Any]):
         """
         Set the opacity of the original point clouds in the figure, not the cuts.
         """
-        self._view.set_opacity(change['new'])
+        self._fig.set_opacity(change['new'])
 
     def toggle_visibility(self):
         """
         Toggle the visibility of the control buttons for making the cuts.
         """
         self.layout.display = None if self.layout.display == 'none' else 'none'
-
-    def toggle_border_visibility(self, change: Dict[str, Any]):
-        """
-        Toggle the visibility of the borders of the cuts.
-        """
-        for cut in self.cuts:
-            cut.toggle_border(change['new'])
-
-    def change_operation(self, change: Dict[str, Any]):
-        """
-        Change the operation to combine multiple cuts.
-        """
-        self._operation = change['new'].lower()
-        self.update_state()
-
-    def update_state(self):
-        """
-        Update the state, combining all the active cuts, using the selected binary
-        operation. The resulting selection is then used to either create or update a
-        second point cloud which is included in the scene.
-        The original point cloud is then set to be semi-transparent.
-        When the position/range of a cut is changed, this function is called via a
-        debounce mechanism to avoid updating the cloud too often. Only the outlines of
-        the cuts are moved in real time, which is cheap.
-        """
-        for nodes in self._nodes.values():
-            self._view.remove(nodes['slice'].id)
-            nodes['slice'].remove()
-        self._nodes.clear()
-
-        visible_cuts = [cut for cut in self.cuts if cut.visible]
-        if not visible_cuts:
-            return
-
-        for n in self._original_nodes:
-            da = n.request_data()
-            selections = []
-            for cut in visible_cuts:
-                xmin, xmax = cut.range
-                selections.append(
-                    (da.coords[cut.dim] > xmin) & (da.coords[cut.dim] < xmax)
-                )
-            selection = reduce(
-                lambda x, y: OPERATIONS[self._operation](x, y), selections
-            )
-            if selection.sum().value > 0:
-                if n.id not in self._nodes:
-                    select_node = Node(selection)
-                    self._nodes[n.id] = {
-                        'select': select_node,
-                        'slice': Node(lambda da, s: da[s], da=n, s=select_node),
-                    }
-                    self._nodes[n.id]['slice'].add_view(self._view)
-                else:
-                    self._nodes[n.id]['select'].func = lambda: selection
-                self._nodes[n.id]['select'].notify_children("")
-
-    @debounce(0.3)
-    def throttled_update(self):
-        self.update_state()
